lambdas/report-stats/requirements.txt
boto3==1.35.36
psycopg2-binary==2.9.9